+++ packages/vite/src/node/ssr/__tests__/ssrTransform.spec.ts
@@ -104,9 +104,9 @@ test('export * from', async () => {
   ).toMatchInlineSnapshot(`
     "const __vite_ssr_import_0__ = await __vite_ssr_import__("vue");
     __vite_ssr_exportAll__(__vite_ssr_import_0__);
+
     const __vite_ssr_import_1__ = await __vite_ssr_import__("react");
     __vite_ssr_exportAll__(__vite_ssr_import_1__);
-
     "
   `)
 })
@@ -247,6 +247,59 @@ test('do not rewrite when function declaration is in scope', async () => {
     function A(){ function fn() {}; return { fn }; }"
   `)
   expect(result?.deps).toEqual(['vue'])
+})
+
+// #16452
+test('do not rewrite when function expression is in scope', async () => {
+  const result = await ssrTransformSimple(
+    `import {fn} from './vue';var a = function() { return function fn() { console.log(fn) } }`,
+  )
+  expect(result?.code).toMatchInlineSnapshot(`
+    "const __vite_ssr_import_0__ = await __vite_ssr_import__("./vue", {"importedNames":["fn"]});
+    var a = function() { return function fn() { console.log(fn) } }"
+  `)
+})
+
+// #16452
+test('do not rewrite when function expression is in global scope', async () => {
+  const result = await ssrTransformSimple(
+    `import {fn} from './vue';foo(function fn(a = fn) { console.log(fn) })`,
+  )
+  expect(result?.code).toMatchInlineSnapshot(`
+    "const __vite_ssr_import_0__ = await __vite_ssr_import__("./vue", {"importedNames":["fn"]});
+    foo(function fn(a = fn) { console.log(fn) })"
+  `)
+})
+
+test('do not rewrite when class declaration is in scope', async () => {
+  const result = await ssrTransformSimple(
+    `import { cls } from 'vue';function A(){ class cls {} return { cls }; }`,
+  )
+  expect(result?.code).toMatchInlineSnapshot(`
+    "const __vite_ssr_import_0__ = await __vite_ssr_import__("vue", {"importedNames":["cls"]});
+    function A(){ class cls {} return { cls }; }"
+  `)
+  expect(result?.deps).toEqual(['vue'])
+})
+
+test('do not rewrite when class expression is in scope', async () => {
+  const result = await ssrTransformSimple(
+    `import { cls } from './vue';var a = function() { return class cls { constructor() { console.log(cls) } } }`,
+  )
+  expect(result?.code).toMatchInlineSnapshot(`
+    "const __vite_ssr_import_0__ = await __vite_ssr_import__("./vue", {"importedNames":["cls"]});
+    var a = function() { return class cls { constructor() { console.log(cls) } } }"
+  `)
+})
+
+test('do not rewrite when class expression is in global scope', async () => {
+  const result = await ssrTransformSimple(
+    `import { cls } from './vue';foo(class cls { constructor() { console.log(cls) } })`,
+  )
+  expect(result?.code).toMatchInlineSnapshot(`
+    "const __vite_ssr_import_0__ = await __vite_ssr_import__("./vue", {"importedNames":["cls"]});
+    foo(class cls { constructor() { console.log(cls) } })"
+  `)
 })
 
 test('do not rewrite catch clause', async () => {
@@ -964,14 +1017,14 @@ console.log(foo + 2)
   `),
   ).toMatchInlineSnapshot(`
     "const __vite_ssr_import_0__ = await __vite_ssr_import__("./foo", {"importedNames":["foo"]});
-    const __vite_ssr_import_1__ = await __vite_ssr_import__("./a");
-    __vite_ssr_exportAll__(__vite_ssr_import_1__);
-    const __vite_ssr_import_2__ = await __vite_ssr_import__("./b");
-    __vite_ssr_exportAll__(__vite_ssr_import_2__);
 
     console.log(__vite_ssr_import_0__.foo + 1)
+    const __vite_ssr_import_1__ = await __vite_ssr_import__("./a");
+    __vite_ssr_exportAll__(__vite_ssr_import_1__);
 
 
+    const __vite_ssr_import_2__ = await __vite_ssr_import__("./b");
+    __vite_ssr_exportAll__(__vite_ssr_import_2__);
 
     console.log(__vite_ssr_import_0__.foo + 2)
       "