+++ playground/optimize-deps/dep-incompatible/index.js
@@ -1,0 +1,3 @@
+const subUrl = new URL('./sub.js', import.meta.url)
+
+export default () => import(subUrl)