+++ playground/resolve-config/__tests__/resolve-config.spec.ts
@@ -13,7 +13,7 @@ const build = (configName: string) => {
 const getDistFile = (configName: string, extension: string) => {
   return fs.readFileSync(
     fromTestDir(`${configName}/dist/index.${extension}`),
-    'utf8',
+    'utf-8',
   )
 }